backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1
pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
